requirements.txt
requests==2.31.0
pydantic==1.10.9
python-dotenv==1.0.0
Pillow==9.5.0
redis==3.5.3
